show data after loading progressBar
