+++ app/src/main/java/com/example/epulapp/beersapplication/dummy/DummyContent.java
@@ -7,7 +7,7 @@ import com.example.epulapp.beersapplication.Model.Beer;
 
 import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
+
 import java.util.Map;
 
 import retrofit2.Call;
@@ -24,42 +24,9 @@ public class DummyContent {
 
     public static final Map<String, Beer> ITEM_MAP = new HashMap<String, Beer>();
 
-    private static final int COUNT = 25;
-    private static final String BASE_URL = "https://api.punkapi.com/v2/";
+    public static Boolean LOADED = false;
 
     static {
-
-        Retrofit retrofit= new Retrofit.Builder()
-                .baseUrl(BASE_URL)
-                .addConverterFactory(GsonConverterFactory.create())
-                .build();
-
-        BiereAPI beerAPI = retrofit.create(BiereAPI.class);
-        Call<ArrayList<Beer>> call = beerAPI.getBeers();
-
-        call.enqueue(new Callback<ArrayList<Beer>>() {
-             @Override
-             public void onResponse(Call<ArrayList<Beer>> call, Response<ArrayList<Beer>> response) {
-                 Log.d("response : ",response.toString());
-                 try {
-                     ArrayList<Beer> beers = response.body();
-                     // Add some sample items.
-                     for (Beer biere: beers) {
-                         Log.d("Biere : ",biere.toString());
-                         ITEMS.add(biere);
-                         ITEM_MAP.put(String.valueOf(biere.getId()), biere);
-                     }
-
-                 } catch (Error e){
-                     Log.e("E : ","error "+e);
-                 }
-             }
-
-             @Override
-             public void onFailure(Call<ArrayList<Beer>> call, Throwable t) {
-                Log.e("E : ","error "+t.getMessage());
-             }
-         });
     }
 
 }
