+++ app/src/main/res/layout/beer_list_content.xml
@@ -4,6 +4,7 @@
     android:layout_height="wrap_content"
     android:orientation="horizontal">
 
+
     <ImageView
         android:id="@+id/image_url"
         android:src="@drawable/ic_launcher_foreground"
